refactor: make the RegisterRoutes component just a helper function
