--- packages/api-reference/playground/multipage/src/package/components/RegisterRouter/index.ts
@@ -1,1 +1,0 @@
-export { default as RegisterRouter } from './RegisterRouter.vue'
